usb_audio_raw: use I2S mode instead of SPI
I2S mode is a lot more flexible when it comes to bitrates, so I can make
an integer decimation factor (in this case, 48k samples per second * 24
seemed to be about the best I could do with the rate I can stuff bits
across USB).  This flexibility comes at the cost of complexity ... this
should be 1152000 bits per second, but I'm only receiving 1015808bps...
not sure where I've miscalculated.

The resulting low-pass-filter-and-resampled audio is much cleaner now
that it's an integer decimation, but I sound like a chipmunk because my
sample rate is off by 13%.