led_loop: repeat the experiment on the stm32f767
I can't use PD9 on the Nucleo-144 board: that's connected to the
ST-Link's serial port on this board.  Moved it to PD7, which is
conveniently at the top-end of one of the sockets on the board.

I'm a little disappointed in performance, this took about 250ns to
respond to the switch press.