Cargo.toml: switch back to opt-mode=z
I don't know where the 2 came from.  At 2 and 3, the wait loops are
unrolled a couple times, which I suppose is sub-ideal branching
behavior.  With z, it's only a handful of assembly instructions to check
the button press and set gpiod.moder:

       0x080002c2 <+178>:   ldr     r0, [r4, #16]
       0x080002c4 <+180>:   lsls    r0, r0, #18
       0x080002c6 <+182>:   bpl.n   0x80002c2 <led_loop::__cortex_m_rt_main+178>
       0x080002c8 <+184>:   movs    r0, #0
       0x080002ca <+186>:   str.w   r0, [r4, #1024] ; 0x400
    => 0x080002ce <+190>:   b.n     0x80002ce <led_loop::__cortex_m_rt_main+190>

This gets it down to about 200ns at 168MHz, and 180ns at 216MHz.