audio: allow controlling the peak-to-peak voltage
I'm not sure if DAC_VOLTAGE is quite right -- my scope's Vpp measurement
seems to read a little higher, but there's also a lot of noise.  VDDA
comes from the 3V rail on the STM32F407G-DISC1, and with this, at least
the Vrms measurement seems to agree with what I'd expect.