echo_pedal: wait for the whole buffer worth
End of Conversion was simply a single ADC sample, not the whole DMA
transaction full of 10 samples we were expecting.

I debugged this by having a GPIO toggle on every iteration of the main
loop, and discovered that it was occasionally (but not reliably!)
toggling at 480k events/sec, which was far too fast!