echo_pedal: remove unnecessary register changes
Obviously we don't need to set the DMA stream's enable to 0 (since we're
waiting on it to become 0), and likewise we don't need to wait a second
time for it to become 0.

ADON doesn't need flipped here, just the ADC DMA needs to be turned off
and back on again.