echo pedal: revisit original design, +TL072
Subjectively, the LM324 was a bit noisier but the bulk of the noise is
still coming out of the STM32's DAC ... remove the connection from PA4
to RV1, and it is immediately silenced, and seems to work fine as a 1:1
buffer for the guitar signal.

The buffer on the input side was added because I was afraid I was
loading my pickups too much and causing attenuation, but that seemed to
make no difference either.