echo pedal: replace STM32 DAC with the CS43L22
This pretty much yanked most of the code from the original usage in the
early days of the signal generator, and reverts the ADC buffer to being
(a) not averaged, and (b) handled in a busy loop, so I don't have to
worry about differences between the clock domains -- it gets slightly
less than 48kHz for the sample clock since MCLK has to be 256x the
sample rate.

This does result in the ADC-captured waveform appearing at the DAC,
though attenuated by a factor of about 12.