echo_pedal: use left alignment from the ADC
the DAC is a 16-bit DAC, so we want our ADC samples to go in the
most-significant 12 bits!

This gets us an output signal that is 800ms delayed, and about the same
amplitude as the input signal.  The DAC does at least sorta block DC and
output near ground, but nothing a capacitor can't fix.