twelve_mhz: actually output 12MHz
This uses 167.5MHz as a hack -- stm32f4xx-hal tries to choose the
largest PLL input it can get away with, but since the PLL output divider
can only go up to 7, we need a 1MHz PLL source clock.

Technically this is still running the PLL's VCO at 60MHz, which does not
meet the 100MHz minimum value.  But it seems to be outputting 12.0007MHz
per the scope's measurement, though the low half of the clock cycle is a
bit shorter than desired yet.