echo_pedal: make sure #CSB is held high
The codec requires a high pulse after each command, and previously it
was only being set high for long enough for a few instructions to
execute -- not even long enough for it to register much voltage change
on the oscilloscope.

"How could this possibly have worked before?"  I have no idea.  My
experiment looked like it worked, but now it reliably actually executes
the power-up command after the reset command.