Enable I2S2 as a slave, indicate events with GPIO
This gets me an edge every time TXE reports an event available, and
another GPIO so I can see what CHSIDE thinks on the scope, so I can
compare it with the LRCLK from the codec.

This is not promising, because I get edges on the TXE-indication that
just run continuously, at the rate of 16 bits / 12MHz (i.e. every 16
bits that are clocked in, even if they're not the most significant 16
bits on a channel).

There is an errata about when to enable the I2S peripheral in slave
mode, though.