Heed the I2S slave errata
Wait for the WS line to go low before actually enabling the I2S
peripheral, like the errata says to do.

This actually works partially as expected!

  * The first TXE event happens immediately (expected, because the
    buffer is completely empty on boot), then
  * The second TXE event happens a while later, shortly after the L/R
    channel clock finally goes high

But it's all downhill from there: the I2S peripheral does not *ever*
wait again for an edge of a L/R clock -- the GPIO that I toggle on TXE
does not have any phase relationship with the L/R clock and bounces all
around in the time domain.